rework task running
