+++ package.json
@@ -8,14 +8,13 @@
     "start": "node deploy",
     "test": "test/run.sh",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
-    "update": "npm-check-updates -u",
+    "ncu": "npm-check-updates -u",
     "webpack": "webpack -p",
     "watch": "webpack --watch",
     "webpack-dev-server": "webpack-dev-server",
     "mocha": "mocha",
     "maven": "mvn -q -f java_parser/pom.xml clean install -U",
     "build": "npm install && npm run maven && npm run webpack",
-    "task": "node runner/runner",
     "kube": "kubectl get rc -o name --selector tier=backend | cut -d '/' -f2 | xargs -n1 kubectl rolling-update --image=yasp/yasp:latest",
     "rediskeys": "redis-cli keys '*' | cut -d':' -f1 | sort | uniq -c",
     "resetpicks": "redis-cli keys 'picks_*' | xargs redis-cli del"
@@ -44,8 +43,6 @@
     "d3": "^3.5.16",
     "datatables": "^1.10.9",
     "datatables-bootstrap3-plugin": "^0.4.0",
-    "datatables.net": "^1.10.11",
-    "datatables.net-bs": "^1.10.11",
     "dota2": "^3.0.0",
     "dota2-minimap-hero-sprites": "^2.0.0",
     "dotenv": "^2.0.0",
